% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_points}
\alias{map_points}
\title{Automated mapping of point locations}
\usage{
map_points(data, maptype = "toner-lite", alpha = 0.5)
}
\arguments{
\item{data}{A dataframe with columns: longitude, latitude.}

\item{maptype}{A Stamen maptype: terrain, terrain-background,
terrain-labels, terrain-lines, toner, toner-2010, toner-2011,
toner-background, toner-hybrid, toner-labels, toner-lines, toner-lite,
 or watercolor.}

\item{alpha}{A transparency value between 0 and 1.}
}
\value{
A plot of location data on a map.
}
\description{
\code{map_points} creates a map of longitude-latitude loction data.
}
\details{
Point location data are mapped over a Stamen basemap. Overplotting
is addressed to some degree with transparence, by default alpha = 0.5.
}
\examples{
map_points(patricks_plants)

boer4 <- filter(patricks_plants, plant_id == "BOER4") # Requires library(dplyr)
map_points(boer4)
}
