% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_presence_absence}
\alias{map_presence_absence}
\title{Automated mapping of presence-absence data}
\usage{
map_presence_absence(df, maptype = "toner-lite", alpha = 0.5)
}
\arguments{
\item{maptype}{A Stamen maptype: terrain, terrain-background,
terrain-labels, terrain-lines, toner, toner-2010, toner-2011,
toner-background, toner-hybrid, toner-labels, toner-lines, toner-lite,
 or watercolor.}

\item{alpha}{A transparency value between 0 and 1.}

\item{data}{A dataframe with columns: presence, latitude, longitude.}
}
\value{
A plot of location data on a map.
}
\description{
\code{map_presence_absence} creates a map of presence-absence data.
}
\details{
Presence-absence data are mapped over a Stamen basemap. Overplotting
is addressed to some degree with transparence, by default alpha = 0.5.
Coordinate datum is assumed to be WGS84.
}
\examples{
boer4 <- presence_absence("BOER4", patricks_plants)
map_presence_absence(boer4)
}
