% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presence_absence.R
\name{presence_absence}
\alias{presence_absence}
\title{Convert presence-only data to presence-absence for a species}
\usage{
presence_absence(plant_id, data = patricks_plants)
}
\arguments{
\item{plant_id}{A character string corresponding to a plant_id}

\item{data}{A presence-only data frame.}
}
\value{
A data frame (tibble) with presence-absence data
}
\description{
\code{presence_absence} converts from presence-only to presence-absence.
}
\details{
The presence-absence may be infered from Patrick's presence-only data.
This function takes a presence-only data frame and a species name and
returns a presence-absence data frame.
}
\examples{
presence_absence("BOER", patricks_plants)
}
