% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_patricks_xlsx.R
\name{read_patricks_xlsx}
\alias{read_patricks_xlsx}
\title{Read Patrick's xlsx file}
\usage{
read_patricks_xlsx(path, coding = "presence-only")
}
\arguments{
\item{path}{Path to xlsx file.}

\item{coding}{Code occurance data as "presence-only" or "presence-absence".}
}
\value{
A data frame (tibble) of plants observed at sites.
}
\description{
\code{read_patricks_xlsx} returns a data frame after parsing Patrick's xlsx file.
}
\details{
Reads and parses Patrick's xlsx file of plant obersvations and metadata into a data frame.
}
\examples{
read_patricks_xlsx("path/to/file.xlsx")
}
